% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate}
\alias{aggregate}
\title{Aggregate morphological profiling data per well.}
\usage{
aggregate(sqlite_file, output_file)
}
\arguments{
\item{sqlite_file}{SQLite data base storing morphological profiles.}

\item{output_file}{Output file for aggregated profiles.}
}
\description{
\code{aggregate} helps you say something.
}
