% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate}
\alias{annotate}
\title{Add plate and well metadata.}
\usage{
annotate(batch_id, plate_id, cell_id = NULL, external_metadata = NULL,
  format_broad_cmap = FALSE, output = NULL,
  perturbation_mode = "chemical", workspace_dir = ".")
}
\arguments{
\item{batch_id}{Batch ID.}

\item{plate_id}{Plate ID.}

\item{cell_id}{Optional cell ID. default: \code{NULL}.}

\item{external_metadata}{Optional external metadata to join with, as a JSON file. default: \code{NULL}.}

\item{format_broad_cmap}{Add columns for compatibility with Broad CMap naming conventions. default: \code{FALSE}.}

\item{output}{Output file (.CSV) for annotated data. If \code{NULL}, writes to \code{workspace_dir/backend/batch_id/plate_id/plate_id_augmented.csv}. default: \code{NULL}.}

\item{perturbation_mode}{Perturbation mode. Must be one of \code{"chemical"} or \code{"genetic"}. default: \code{"chemical"}.}

\item{workspace_dir}{Root directory containing backend and metadata subdirectories. Can be relative or absolute. default: \code{"."}.}
}
\description{
\code{annotate} adds plate and well metadata, as well as optional external
metadata. It assumes the following files exist:
- \code{<workspace_dir>/metadata/barcode_platemap.csv} which maps the plate
barcodes to the corresponding platemaps,
- For each platemap, a file
\code{<workspace_dir>/metadata/platemap/<plate_map_name>.txt} which
contains the metadata for each well position
}
