% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{asinh_transform}
\alias{asinh_transform}
\title{Hyperbolic Arcsin Transform for Cytometry Data}
\usage{
asinh_transform(x)
}
\arguments{
\item{x}{vector to transform}
}
\value{
The transformed vector.
}
\description{
\code{asinh_transform} applies the standard transform applied to
    channels of CyToF or flow cytometry data when comparing channels
    directly.
}
\details{
Performs the \code{asinh(x / 5)} transformation to the input data.
}
\references{
Garner, Duane L., et al. "Fluorometric assessments of
    mitochondrial function and viability in cryopreserved bovine
    spermatozoa." Biology of Reproduction 57.6 (1997): 1401-1406.
}
