% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\docType{data}
\name{emd_downsample_rows_default}
\alias{emd_downsample_rows_default}
\alias{emd_comparison_runs_default}
\title{?}
\format{An object of class \code{integer} of length 1.}
\usage{
emd_downsample_rows_default

emd_comparison_runs_default
}
\description{
?

?
}
\keyword{datasets}
