% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLogicle.R
\name{estimateLogicles}
\alias{estimateLogicles}
\title{Estimation of linearization width for Logicle transformation across a set 
of FCS files}
\usage{
estimateLogicles(files)
}
\arguments{
\item{files}{List of FCS file paths}
}
\value{
A dataframe with estimated linearization width for every channel listed
in $SPILL slot
}
\description{
\code{estimateLogicles} reads FCS and compensates an FCS file
and subsequently estimates the linearization width parameter for a logicle 
transformation. Note that reading all FCS files can take some time for large 
sets of files.
}
