% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gateDoublets.R
\name{gateDoublets}
\alias{gateDoublets}
\title{Automated doublet gating}
\usage{
gateDoublets(ff)
}
\arguments{
\item{ff}{A flowframe}

\item{channel1}{Name of FSC-A channel (default = "FSC-A")}

\item{channel2}{Name of FSC-H channel (default = "FSC-H")}

\item{predictionLevel}{Value of prediction level used (flowStats)}

\item{nMAD}{Number of median absolute deviations (PeacoQC)}

\item{nBins}{Number of bins (Splines)}

\item{nSD}{Number of standard deviations (Splines)}
}
\value{
A matrix with a binary (0, 1) column for every method indicating
whether a cell is a doublet (1) or singlet (0)
}
\description{
\code{gateDoublets} identifies doublets using three methods
}
