% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_marker_names}
\alias{get_marker_names}
\title{?}
\usage{
get_marker_names(frames, normalize = toupper, rx_replace = c(),
  name_filter = non_pheno_channel_name_patterns,
  content_filter = non_pheno_channel_content_predicates,
  check.spelling = TRUE, check.channel_switches = TRUE)
}
\arguments{
\item{frames}{?}

\item{normalize}{?}

\item{rx_replace}{?}

\item{name_filter}{?}

\item{content_filter}{?}

\item{check.spelling}{?}

\item{check.channel_switches}{?}
}
\value{
?
}
\description{
?
}
\details{
?
}
