% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCutoffclusters.R
\name{getCutoffclusters}
\alias{getCutoffclusters}
\title{Perform metaclustering based on manually identified MFI cutoffs}
\usage{
getCutoffclusters(MFIData, cutoffs)
}
\arguments{
\item{MFIData}{Matrix of MFIs of every marker (column) on every cluster (row)}

\item{cutoffs}{Named list of parameters (corresponding to MFIData columns) and cutoffs}
}
\value{
A named list containing the patterns, cutoffs and metaclustering
}
\description{
\code{getCutoffclusters} assesses the MFIs of a given marker across
a range of clusters. Using a named list of cutoffs, clusters are assigned
as either marker+ or marker-. This labeling is converted into a metaclustering.
}
