% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFCSmetadata.R
\name{getFCSmetadata}
\alias{getFCSmetadata}
\title{Fast metadata extraction from a list of FCS files}
\usage{
getFCSmetadata(
  files,
  commonParams = c("$FIL", "GUID", "EXPORT TIME", "CYTNUM"),
  ignore = NULL
)
}
\arguments{
\item{files}{List of FCS files}

\item{commonParams}{List of common FCS parameters to save}

\item{ignore}{List of parameters from commonParams to ignore}
}
\description{
\code{getFCSmetadata} uses flowCore to read the headers from FCS
files and returns a large dataframe with metadata for all files.

Note that this function can take long for very large sets of files. In that case,
run the function for subsets and concatenate the output!
}
