% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLogicle.R
\name{getLogicle}
\alias{getLogicle}
\title{Get an optimized logicle transformList object based on a set of FCS files}
\usage{
getLogicle(files)
}
\arguments{
\item{files}{List of FCS file paths}

\item{verbose}{Option for printing identified parameters}
}
\value{
A flowCore transformList object
}
\description{
Uses \code{estimateLogicles} to determine the optimal linearization
width for every channel in a set of FCS files and builds a transformList object
using the median width for every channel
}
\seealso{
\code{\link{estimateLogicles}}
}
