% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMFIclusters.R
\name{labelPeaks}
\alias{labelPeaks}
\title{Label peaks in a numerical vector of MFIs across clusters}
\usage{
labelPeaks(values)
}
\arguments{
\item{values}{Vector of MFIs for a marker across clusters}
}
\value{
A named list with cutoffs and labels for all clusters
}
\description{
\code{labelPeaks} assesses the MFIs of a given marker across
a range of clusters. The number of peaks (max 3) are detected and labeled
as either "unimodal" (1 peak), "low/high" (2 peaks) or "low/medium/high".
}
