% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{list_dir_files}
\alias{list_dir_files}
\title{List Files in a Directory}
\usage{
list_dir_files(path)
}
\arguments{
\item{path}{string, the directory to read.}
}
\value{
Character vector containing paths to all files located at the
specified \code{path}.
}
\description{
\code{list_dir_files} lists all files in a directory without
including other directories and using the full filenames.
}
\examples{
## list all files in the current directory
list_dir_files(".")

}
