% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchFCSexports.R
\name{matchFCSexports}
\alias{matchFCSexports}
\title{Match individually exported gated populations to a full FCS file}
\usage{
matchFCSexports(file, exports, labels = NULL, matchVar = "event_ID")
}
\arguments{
\item{file}{The original full FCS file}

\item{exports}{List of FCS files containing exported populations}

\item{labels}{How to label every export in the output dataframe. By default,
uses the full path of every file}

\item{matchVar}{The variable in each FCS file to use for matching}
}
\value{
A dataframe indicating for every event whether its present in each
exported FCS file
}
\description{
\code{matchFCSexports} uses a common variable between a full
FCS file and individually exported FCS files to re-construct a gating hierarchy.
}
