% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\docType{data}
\name{mem_iqr_threshold_default}
\alias{mem_iqr_threshold_default}
\title{?}
\format{An object of class \code{numeric} of length 1.}
\usage{
mem_iqr_threshold_default
}
\description{
?
}
\references{
Diggins, Kirsten E., et al. "Characterizing cell subsets using
marker enrichment modeling." Nature Methods 14.3 (2017): 275-278.
}
\keyword{datasets}
