% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{multi_str_split}
\alias{multi_str_split}
\title{Split Elements of a Character Vector By Multiple Elements}
\usage{
multi_str_split(strs, splits, fixed)
}
\arguments{
\item{strs}{character vector, each element of which is to be split.}

\item{splits}{character vector of fixed strings or PCRE regular expressions
to use for splitting. If \code{splits} has length 0, \code{strs} is split
into single characters. All elements of \code{splits} are used to split
every element of \code{strs}!}

\item{fixed}{logical. If \code{TRUE} match \code{splits} exactly, otherwise
use PCRE regular expressions.}
}
\value{
A list of the same length as \code{strs}, where each element is a
character vector containing the splits of the corresponding element of
\code{strs} according to \code{splits}.
}
\description{
\code{multi_str_split} splits the elements of \code{strs} into substrings
between instances of any element of \code{splits}.
}
\examples{
multi_str_split(c("a_b-c d", "a-b_c d", "123"), c("-", "_", " ", "2"),
    fixed = T)
## [[1]]
## [1] "a" "b" "c" "d"
##
## [[2]]
## [1] "a" "b" "c" "d"
##
## [[3]]
## [1] "1" "3"

}
\seealso{
\code{\link[=strsplit]{strsplit()}} is used to perform the splitting.
}
