% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\docType{data}
\name{is_non_floating_point}
\alias{is_non_floating_point}
\alias{non_pheno_channel_content_predicates}
\title{?}
\format{An object of class \code{list} of length 1.}
\usage{
is_non_floating_point(vec)

non_pheno_channel_content_predicates
}
\arguments{
\item{vec}{?}
}
\value{
?
}
\description{
?

?
}
\keyword{datasets}
