% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\docType{data}
\name{non_pheno_channel_name_patterns}
\alias{non_pheno_channel_name_patterns}
\title{?}
\format{An object of class \code{list} of length 3.}
\usage{
non_pheno_channel_name_patterns
}
\description{
?
}
\keyword{datasets}
