% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{normalize_channels}
\alias{normalize_channels}
\title{?}
\usage{
normalize_channels(frame, desc,
  content_filter = non_pheno_channel_content_predicates,
  normalize_channel_names = toupper,
  channel_name_ops = non_pheno_channel_name_patterns)
}
\arguments{
\item{frame}{?}
}
\value{
?
}
\description{
?
}
