% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{normalize_pheno_channels_dataset}
\alias{normalize_pheno_channels_dataset}
\title{Extract and Transform Phenotype Channels from a CyToF Dataset}
\usage{
normalize_pheno_channels_dataset(pheno_channel_names, pop_list, ref = NULL,
  transform_fun = asinh_transform(5))
}
\arguments{
\item{pheno_channel_names}{character vector without duplicates containing
names of channels to extract from each population. See details for how it
is interpreted.}

\item{pop_list}{list of data frames containing CyToF data as from
\code{\link[=read_files_sorted]{read_files_sorted()}}.}

\item{ref}{a data frame containing CyToF data, as from \code{\link[=read_cyto_file]{read_cyto_file()}}, or
NULL. If NULL, the populations in \code{pop_list} are joined with \code{\link[=rbind]{rbind()}}
after their channel names are converted to \code{pheno_channel_names}.}

\item{transform_fun}{vectorized function to transform the raw channel values,
such as generated by \code{\link[=asinh_transform]{asinh_transform()}}.}
}
\value{
named list with keys \code{pop_list} and \code{ref}.\preformatted{The key `pop_list` contains the list of data frames from the argument
`pop_list` after extracting `pheno_channel_names`. The [names()] of the
argument `pop_list` are reused if the list has them, otherwise the
indices `1:length(pop_list)` are coerced to a character vector and used
as names instead.

The key `ref` contains a data frame with either the reference population
given by the argument `ref` (with the columns matched by
`pheno_channel_names` extracted), or all the events in `pop_list` joined
together with [rbind()] after extracting `pheno_channel_names`.
}
}
\description{
\code{normalize_pheno_channels_dataset} matches a set of canonical
channel names containing phenotype data over a list of populations and
reference population, and applies a specified transformation to the
phenotype channels.
}
\details{
?
}
\seealso{
\code{\link[=read_files_sorted]{read_files_sorted()}} and \code{\link[=read_cyto_file]{read_cyto_file()}} to use as inputs to
\code{pop_list} and \code{ref}, and \code{\link[=asinh_transform]{asinh_transform()}} to use as a transformation
on the raw channel values.
}
