% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs.R
\name{pairwise_emd}
\alias{pairwise_emd}
\title{Compute pairwise EMD of a set of datasets in a viSNE analysis.}
\usage{
pairwise_emd(frames, outfile, max_iterations = 10, visne_axes = c("tSNE1",
  "tSNE2"), verbose = T)
}
\arguments{
\item{frames}{named list of data frames representing CyToF
datasets. \code{\link{process_cyto_dataset}} can be used to produce this.}

\item{outfile}{string naming a file path or an open connection which the
resultant matrix of comparisons is written to with
\code{\link{write.table}}.}

\item{max_iterations}{Positive integer indicating the maximum iterations used
to compute EMD. Changing this argument \emph{typically} does not change
the result very much.}

\item{visne_axes}{character vector indicating the column names containing the
viSNE channels.}

\item{verbose}{logical. Whether to print progress indicators to the console.}
}
\value{
The value of \code{outfile}. \code{outfile} contains both row and
    column names, and should be read back into a matrix for analysis like:
    \code{as.matrix(read.csv(outfile, row.names = 1))}.
}
\description{
\code{pairwise_emd} computes the Earth Mover's Distance (EMD)
    between the viSNE axes of each pair of data frames given, and writes the
    resultant double-precision matrix to a CSV file.
}
\details{
If \code{length(frames) == n} for some positive integer \code{n}, a
    diagonal n x n matrix \code{mat} is created to represent the EMD between
    each pair of datasets at indices \code{i} and \code{j}, where
    \code{mat[i,j] == mat[j,i]} and \code{mat[i,j]} represents the EMD
    between the two.
}
\seealso{
\code{\link{emdist::emdw}} for the underlying EMD implementation.
}
