% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{pairwise_mem_rmsd}
\alias{pairwise_mem_rmsd}
\title{Compute pairwise MEM RMSD of a set of CyToF datasets.}
\usage{
pairwise_mem_rmsd(frames, outfile, markers = NULL, ref_pop = NULL,
  transform_with = asinh_transform, verbose = T)
}
\arguments{
\item{frames}{named list of data frames representing CyToF
datasets. \code{\link{process_cyto_dataset}} can be used to produce this.}

\item{outfile}{string naming a file path or an open connection which the
resultant matrix of comparisons is written to with
\code{\link{write.table}}.}

\item{markers}{character vector of channel names to use for the MEM
calculation, or \code{NULL}. If \code{markers = NULL}, the function will
guess the markers to perform the MEM analysis on. If \code{verbose =
TRUE}, the choice of markers will be printed to the console. If
\code{markers} is non-\code{NULL}, all of the columns indicated should
exist in each input dataset.}

\item{ref_pop}{data frame produced by \code{\link{read_cyto_file}} containing
the reference population to use for the MEM calculation, or
\code{NULL}. If \code{ref_pop = NULL}, each input dataset will be
collapsed (using \code{\link{rbind}}) into a single giant reference
population.}

\item{transform_with}{specifies how the raw channel values should be #}

\item{verbose}{logical. Whether to print progress indicators to the console.}
}
\value{
The value of \code{outfile}. \code{outfile} contains both row and
    column names, and should be read back into a matrix for analysis like:
    \code{as.matrix(read.csv(outfile, row.names = 1))}.
}
\description{
\code{pairwise_mem_rmsd} computes the Root Mean Square Distance
    (RMSD) of the Marker Enrichment Modeling (MEM) score between each each
    pair of data frames given, and writes the resultant double-precision
    matrix to a CSV file.
}
\details{
If \code{length(frames) == n} for some positive integer \code{n}, a
    diagonal n x n matrix \code{mat} is created to represent the MEM RMSD
    between each pair of datasets at indices \code{i} and \code{j}, where
    \code{mat[i,j] == mat[j,i]} and \code{mat[i,j]} represents the MEM RMSD
    between the two.
}
\references{
Diggins, Kirsten E., et al. "Characterizing cell subsets using
    marker enrichment modeling." Nature Methods 14.3 (2017): 275-278.
}
\seealso{
\code{\link{asinh_transform}} is the default transform used to
    compare markers.
}
