% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs.R
\name{plot_pairwise_comparison}
\alias{plot_pairwise_comparison}
\title{Plot Pairwise Comparisons of Cytometry Datasets}
\usage{
plot_pairwise_comparison(matrix_file, color_palette = NULL)
}
\arguments{
\item{matrix_file}{string naming a file path or an open connection which a
comparison matrix is read from. File should be produced by
\code{\link{pairwise_emd}} or \code{\link{pairwise_mem_rmsd}}.}

\item{color_palette}{color palette generated by
\code{\link{grDevices::colorRampPalette}}, or \code{NULL}.}
}
\description{
\code{plot_pairwise_comparison} plots a heatmap of a pairwise
    comparison of datasets with \code{\link{gplots::heatmap.2}}.
}
\seealso{
\code{\link{gplots::heatmap.2}} for the underlying plotting
    function, and \code{\link{grDevices::colorRampPalette}} for color palette
    generation. \code{\link{pairwise_emd}} or \code{\link{pairwise_mem_rmsd}}
    should be used to generate \code{matrix_file}.
}
