% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preselect.R
\name{preselect}
\alias{preselect}
\title{Preselect...}
\usage{
preselect(batch_id, input, operations, replicates = NULL, subset = NULL,
  output_dir = NULL, workspace_dir = ".")
}
\arguments{
\item{batch_id}{Batch ID.}

\item{input}{Test data on which to perform variable selection operations. Must be CSV.}

\item{operations}{List of operations to perform, as strings. Supported operations: \code{"correlation_threshold"}, \code{"variance_threshold"}, \code{"replicate_correlation"}.}

\item{replicates}{Number of replicates to select per plate map. Required for the operation \code{"replicate_correlation"}. default: \code{NULL}.}

\item{subset}{Query to create the training data by subsetting. Must be CSVs. default: \code{NULL}.}

\item{output_dir}{Output directory for preselected feature names. If \code{NULL}, writes to \code{workspace_dir/parameters/batch_id/variable_selection/}. default: \code{NULL}.}

\item{workspace_dir}{Root directory containing backend and metadata subdirectories. Can be relative or absolute. default: \code{"."}.}
}
\description{
\code{preselect}
}
