% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_files_sorted}
\alias{read_files_sorted}
\title{Scan a Directory for CytoF Files, Sort Them, and Read Them Into Memory}
\usage{
read_files_sorted(path = ".", extensions = c("fcs"), split_by = c("_",
  " "), orders = list(), allow_skip = TRUE)
}
\arguments{
\item{path}{string, the directory to read.}

\item{extensions}{character vector containing file extensions to process.}

\item{split_by}{character vector used to split \code{strs} into
components. This is interpreted as either fixed strings or PCRE regular
expressions according to \code{fixed}.}

\item{orders}{list of character vectors indicating a lexicographic ordering
for each component.}

\item{allow_skip}{logical indicating whether to allow reading text files with
one blank line at the top.}
}
\value{
Named list of data frames, where each name is the filename from which
the value at that index is read.
}
\description{
\code{read_files_sorted} reads all files with the given extensions in
the current directory into a named list of data frames.
}
\details{
The directory specified by \code{path} is scanned for files with any of
the given \code{extensions}. The elements of \code{split_by} are interpreted as
\emph{fixed strings} to split the resulting filenames, and \code{orders} are used
to lexicographically sort them with \code{\link[=sort_by_component]{sort_by_component()}}. The resulting
sorted character vector of filenames is then used as the \code{\link[=names]{names()}} for a
named list, which is then populated with each file's contents using
\code{\link[=read_cyto_file]{read_cyto_file()}}.
}
\examples{
sorted_cytof_data <- read_files_sorted(
   path = ".",
   extensions = c("fcs"),
   split_by = c("_", " "),
   orders = list(c(), c("pre", "3wk", "12wk", "6m")))

length(sorted_cytof_data)
## [1] 141
is.data.frame(sorted_cytof_data[[1]])
## [1] TRUE

lapply(sorted_cytof_data[1:2], function (fcs_df) fcs_df[1:4,c(1:2,5:6)])
## $`./29_RCCPBMC_panel4_PD-1+ CD4.fcs`
##    Time Cell_length        140  ICOS-141
## 1 19648          24  0.2647168  1.667185
## 2 20933          17  1.0358380  9.964522
## 3 27051          24 -0.8677309 17.684467
## 4 28387          37  1.2802999 14.215277
##
## $`./29_RCCPBMC_panel4_PD-1+ CD8.fcs`
##    Time Cell_length        140    ICOS-141
## 1 29018          22 -0.9847549 -0.51542509
## 2 30535          31 -0.5153459 -0.65859687
## 3 33917          19 -1.0417140  3.72425652
## 4 42929          27  1.7391516  0.03845136

}
\seealso{
\code{\link[=list_dir_files]{list_dir_files()}} is used to scan the directory,
\code{\link[=sort_by_component]{sort_by_component()}} sorts file paths, and \code{\link[=read_cyto_file]{read_cyto_file()}} is called
to interpret each file's contents into a data frame.
}
