% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularize.R
\name{regularize}
\alias{regularize}
\title{Regularize backend file}
\usage{
regularize(input_file, output_file)
}
\arguments{
\item{input_file}{Input file. CSV or SQLite only.}

\item{output_file}{Output file. Same format as input_file.}
}
\description{
\code{regularize} regularizes backend file.
}
