% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{rotate_matrix_ccw}
\alias{rotate_matrix_ccw}
\title{Rotate a Matrix Counter-Clockwise}
\usage{
rotate_matrix_ccw(m)
}
\arguments{
\item{mat}{Any matrix.}
}
\value{
Return an n x m matrix obtained by rotating \code{mat} counter-clockwise,
if \code{mat} is an m x n matrix.
}
\description{
\code{rotate_matrix_ccw} performs a counter-clockwise rotation of the
input matrix.
}
\examples{
mat <- matrix(1:8, ncol = 2)
mat
##      [,1] [,2]
## [1,]    1    5
## [2,]    2    6
## [3,]    3    7
## [4,]    4    8
rotate_matrix_ccw(mat)
##      [,1] [,2] [,3] [,4]
## [1,]    5    6    7    8
## [2,]    1    2    3    4

}
