% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample}
\alias{sample}
\title{**DESCRIPTION**}
\usage{
sample(batch_id, pattern, output, replicates = NULL, workspace_dir = ".")
}
\arguments{
\item{batch_id}{Batch ID.}

\item{pattern}{Regular expression - only CSVs.}

\item{output}{Output file - either CSV or RDS.}

\item{replicates}{Number of replicates to select per plate map. Optional. default: \code{NULL}.}

\item{workspace_dir}{Root directory containing backend and metadata subdirectories. Can be relative or absolute. default: \code{"."}.}
}
\description{
\code{sample} **DESCRIPTION**
}
