% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{timed_execute}
\alias{timed_execute}
\title{?}
\usage{
timed_execute(expr, digits = 3, time_type = 3, env = parent.frame())
}
\arguments{
\item{expr}{?}

\item{digits}{?}

\item{time_type}{?}

\item{env}{?}

\item{drop_time}{?}
}
\description{
?
}
