% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gct.R
\name{write_gct}
\alias{write_gct}
\title{Write CellProfiler data to gct file.}
\usage{
write_gct(
  x,
  path,
  channels = NULL,
  feature_regex = "^Nuclei_|^Cells_|^Cytoplasm_",
  create_row_annotations = TRUE
)
}
\value{
The input \code{x}, invisibly.
}
\description{
@param x ...
 @param path ...
 @param channels ...
}
