% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate}
\alias{aggregate}
\title{Aggregate data based on given grouping.}
\usage{
aggregate(sqlite_file, output_file, compartments = c("cells",
  "cytoplasm", "nuclei"), operation = "mean",
  strata = c("Metadata_Plate", "Metadata_Well"), variables = "all",
  univariate = TRUE)
}
\arguments{
\item{sqlite_file}{SQLite database storing morphological profiles.}

\item{output_file}{Output file for storing aggregated profiles.}

\item{compartments}{Optional character vector specifying cellular compartments. default: \code{c("cells", "cytoplasm", "nuclei")}.}

\item{operation}{Optional character string specifying method for aggregation, e.g. \code{"mean"}, \code{"median"}, \code{"mean+sd"}. See \link[cytominer]{aggregate}. default: \code{"mean"}.}

\item{strata}{Character vector specifying grouping variables for aggregation. default: \code{c("Metadata_Plate", "Metadata_Well")}.}

\item{variables}{Optional character vector specifying observation variables. default: \code{"all"}.}

\item{univariate}{Optional boolean specifying whether aggregation function is univariate (vs multivariate). default: \code{"TRUE"}.}
}
\description{
\code{aggregate} aggregates data based on the specified aggregation method.
}
