% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate}
\alias{aggregate}
\title{Aggregate data based on given grouping.}
\usage{
aggregate(sqlite_file, output_file, compartments = c("cells",
  "cytoplasm", "nuclei"), operation = "mean",
  strata = c("Metadata_Plate", "Metadata_Well"), variables = "all")
}
\arguments{
\item{sqlite_file}{SQLite database storing morphological profiles.}

\item{output_file}{Output file for storing aggregated profiles.}

\item{compartments}{optional character vector specifying cellular compartments. default \code{c("cells", "cytoplasm", "nuclei")}.}

\item{operation}{optional character string specifying method for aggregation, e.g. \code{"mean"}, \code{"median"}, \code{"mean+sd"}. default \code{"mean"}.}

\item{strata}{character vector specifying grouping variables for aggregation. default \code{c("Metadata_Plate", "Metadata_Well")}.}

\item{variables}{optional character vector specifying observation variables. default \code{"all"}.}
}
\description{
\code{aggregate} aggregates data based on the specified aggregation method.
}
