% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate}
\alias{aggregate}
\title{Aggregate data based on given grouping.}
\usage{
aggregate(output_file, sqlite_file, strata = c("Image_Metadata_Plate",
  "Image_Metadata_Well"), operation = "median")
}
\arguments{
\item{output_file}{Output file for storing aggregated profiles.}

\item{sqlite_file}{SQLite database storing morphological profiles.}

\item{strata}{character vector specifying grouping variables for aggregation. default \code{c("Image_Metadata_Plate", "Image_Metadata_Well")}.}

\item{operation}{optional character string specifying method for aggregation. This must be one of the strings \code{"mean"}, \code{"median"}, \code{"mean+sd"}. default \code{"median"}.}
}
\description{
\code{aggregate} aggregates data based on the specified aggregation method.
}
