% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{calc_mem}
\alias{calc_mem}
\title{?}
\usage{
calc_mem(pop_list, ref, IQRthresh = mem_iqr_threshold_default,
  scale_limit = NULL)
}
\arguments{
\item{pop_list}{?}

\item{ref}{?}

\item{IQRthresh}{?}

\item{scale_limit}{?}
}
\value{
?
}
\description{

}
\details{
?
}
\references{
Diggins, Kirsten E., et al. "Characterizing cell subsets using
    marker enrichment modeling." Nature Methods 14.3 (2017): 275-278.
}
