% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{calc_mem}
\alias{calc_mem}
\title{Compute MEM Scores Over Populations in a Dataset}
\usage{
calc_mem(pop_list, ref, IQRthresh = 0.5, scale_limit = NULL)
}
\arguments{
\item{pop_list}{list of floating-point matrices or data frames, each with the
same column names in the same order. Each column represents \emph{transformed}
expression levels of the named phenotype.}

\item{ref}{floating-point matrix or data frame with the same column names in
the same order as each element of \code{pop_list}, with entries representing
\emph{transformed} expression levels. Used as reference population.}

\item{IQRthresh}{positive scalar floating-point, used as lower bound for IQR
calculations (see MEM paper).}

\item{scale_limit}{positive scalar floating-point or \code{NULL}. If not \code{NULL},
MEM values are scaled so the greatest maximum absolute MEM value is equal
to \code{scale_limit}, otherwise no scaling occurs.}
}
\value{
matrix with column names as in \code{pop_list} and \code{ref}, row names equal
to the names of \code{pop_list} (if given). Each row contains the MEM values
of a population in \code{pop_list} for each phenotype channel, in the same
order as given.
}
\description{
\code{calc_mem} computes the Marker Enrichment Modelling (MEM) score
of each population given against the provided reference population.
}
\details{
Note that the values of entries in \code{ref} and the populations of
\code{pop_list} are assumed to already be transformed (typically an arcsinh
transform, as in \code{\link[=asinh_transform]{asinh_transform()}}).\preformatted{To obtain a distance matrix of the Root Mean Square Distance (RMSD)
betwen each population's MEM scores, the [dist()] function can be used,
and converted to a matrix with [as.matrix()].
}
}
\examples{
## make some example data (values are not representative of real data)
## inputs to `calc_mem` should be transformed already (e.g. the below
## could represent an arcsinh-transformed dataset)
a <- data.frame(CCR5 = rnorm(100, 0, 5), CD3 = rnorm(100, 0, 5))
b <- data.frame(CCR5 = rnorm(150, 0, 5), CD3 = rnorm(150, 0, 5))
c <- data.frame(CCR5 = rnorm(150, 0, 5), CD3 = rnorm(150, 0, 5))
## example reference
ref <- data.frame(CCR5 = rnorm(300, 0, 5), CD3 = rnorm(300, 0, 5))

## calculate MEM with minimum IQR of 0.5, scaled so maximum absolute
## magnitude is equal to 10 (as suggested in MEM paper)
mems <- calc_mem(list("pop A" = a, "pop B" = b),
         ref,
         IQRthresh = 0.5, scale_limit = 10)
mems
##            CCR5       CD3
## pop A -2.446713 -1.085913
## pop B -1.304841 -1.620452
## pop C  3.201968 10.000000

## make a distance matrix of the RMSD between populations
as.matrix(dist(mems))
##           pop A     pop B    pop C
## pop A  0.000000  1.260795 12.44207
## pop B  1.260795  0.000000 12.46380
## pop C 12.442068 12.463797  0.00000

}
\references{
Diggins, Kirsten E., et al. "Characterizing cell subsets using
marker enrichment modeling." Nature Methods 14.3 (2017): 275-278.
}
\seealso{
\code{\link[=normalize_pheno_channels_dataset]{normalize_pheno_channels_dataset()}} can be used to generate
a suitable \code{pop_list} and \code{ref} from a CyToF dataset, applying
transformations and normalizing phenotype channel names.
}
