% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\docType{data}
\name{numeric_channel_name}
\alias{numeric_channel_name}
\alias{sne_channel_name}
\alias{non_pheno_channel_name_patterns}
\title{?}
\format{An object of class \code{character} of length 1.}
\usage{
numeric_channel_name

sne_channel_name

non_pheno_channel_name_patterns
}
\description{
?

?

?
}
\keyword{datasets}
