% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize profiles based on given control subset}
\usage{
normalize(batch_id, plate_id, subset, backend_directory = file.path("..",
  "..", "backend"), tmpdir = "/tmp", operation = "robustize",
  sample_single_cell = FALSE)
}
\arguments{
\item{batch_id}{batch ID given as directory name}

\item{plate_id}{plate ID given as the directory}

\item{subset}{Query to define the control poulation used for normalization \code{("Metadata_line_num == 5")}}

\item{backend_directory}{Top level directory of the experiment, default: \code{"../../backend/"},}

\item{tmpdir}{Temporary directory, default \code{"/tmp"}.}

\item{operation}{Method used for normalization, default: \code{"robustize"}. See cytominer::normalize}

\item{sample_single_cell}{Use single cell data for sampling, instead of per-well profiles, default \code{FALSE}.}
}
\description{
Normalize profiles based on given control subset
}
