% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{normalize_pheno_channels_dataset}
\alias{normalize_pheno_channels_dataset}
\title{?}
\usage{
normalize_pheno_channels_dataset(pheno_channel_names, pop_list, ref = NULL,
  transform_fun = asinh_transform)
}
\arguments{
\item{pop_list}{?}

\item{ref}{?}

\item{transform_fun}{?}

\item{content_filter}{?}

\item{modify_colnames_fun}{?}

\item{channel_name_ops}{?}
}
\value{
?
}
\description{
?
}
