% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{plot_pairwise_comparison}
\alias{plot_pairwise_comparison}
\title{Plot Pairwise Comparisons of Cytometry Datasets}
\usage{
plot_pairwise_comparison(mat, pcnt_similarity_scale = FALSE,
  with_dendrograms = FALSE, ...)
}
\arguments{
\item{mat}{numeric matrix, such as produced by \code{\link[=pairwise_emd]{pairwise_emd()}} or
\code{\link[=pairwise_mem_rmsd]{pairwise_mem_rmsd()}}. \code{\link[=rotate_matrix_ccw]{rotate_matrix_ccw()}} can be used to rotate \code{mat}
as desired. \emph{must} have \code{\link[=dimnames]{dimnames()}}!}

\item{pcnt_similarity_scale}{logical, whether to read \code{mat} as a distance
matrix and scale its values from 0-100, with 0 for 0 entries in \code{mat} and
100 for the greatest entry in \code{mat}, scaled linearly for all others.}

\item{with_dendrograms}{logical, whether to use dendrograms on the output.}

\item{...}{additional arguments to pass to \code{\link[gplots:heatmap.2]{gplots::heatmap.2()}}. These take
precedence over any arguments this function tries to pass.}
}
\description{
\code{plot_pairwise_comparison} plots a heatmap of a pairwise
comparison of datasets with \code{\link[gplots:heatmap.2]{gplots::heatmap.2()}}.
}
\details{
\code{mat} can be any numeric matrix, but this function is specifically
designed for the case when \code{mat} is a distance matrix (symmetric, all
values nonnegative). \code{pcnt_similarity_scale = TRUE} creates a
"similarity matrix" from \code{mat}, scaling its values linearly from 0-100,
using 100 for 0 entries of \code{mat} and 0 for the greatest entry of \code{mat}.
}
\examples{
mat <- as.matrix(dist(rnorm(5)))
## row/colnames are required for this function
dimnames(mat) <- list(c("uno", "dos", "tres", "cuatro", "cinco"),
                      c("uno", "dos", "tres", "cuatro", "cinco"))
mat
##              uno       dos       tres     cuatro      cinco
## uno    0.0000000 0.3647115 1.14159784 1.05391169 1.17646111
## dos    0.3647115 0.0000000 0.77688632 0.68920017 0.81174959
## tres   1.1415978 0.7768863 0.00000000 0.08768615 0.03486327
## cuatro 1.0539117 0.6892002 0.08768615 0.00000000 0.12254942
## cinco  1.1764611 0.8117496 0.03486327 0.12254942 0.00000000

## no dendrograms, no scaling (default)
plot_pairwise_comparison(mat)
## (R graphics device)

## rotated
plot_pairwise_comparison(rotate_matrix_ccw(mat))
## (R graphics device)

## add similarity scaling and dendrograms
plot_pairwise_comparison(
    mat,
    pcnt_similarity_scale = TRUE, with_dendrograms = TRUE,
    cexRow = 1.5, cexCol = 1.5, margin = c(5, 5))
## (R graphics device)

}
\seealso{
\code{\link[gplots:heatmap.2]{gplots::heatmap.2()}} for the underlying plotting
function, and \code{\link[grDevices:colorRampPalette]{grDevices::colorRampPalette()}} for color palette
generation. \code{\link[=pairwise_emd]{pairwise_emd()}} or \code{\link[=pairwise_mem_rmsd]{pairwise_mem_rmsd()}}
should be used to generate \code{matrix_file}. \code{\link[=rotate_matrix_ccw]{rotate_matrix_ccw()}} can be
used to rotate the matrix \code{mat} as desired.
}
